// apps/cofepris/src/app/application/tramites/260910/models/mercancia.model.ts
export interface DestinatarioImitar {
  /** Tipo de persona (física o moral). */
  tipoPersona: string;

  /** Indicación para modificar el RFC del destinatario. */
  modificarRFC: string;

  /** Denominación social del destinatario. */
  denominacion: string;

  /** Identificador del país asociado al domicilio. */
  domicilioPais: number;

  /** Estado asociado al domicilio del destinatario. */
  domicilioEstado: string;

  /** Municipio asociado al domicilio del destinatario. */
  domicilioMunicipio: string;

  /** Localidad asociada al domicilio del destinatario. */
  domicilioLocalidad: string;

  /** Código postal del domicilio del destinatario. */
  domicilioCodigo: string;

  /** Colonia asociada al domicilio del destinatario. */
  domicilioColonia: string;

  /** Calle del domicilio del destinatario. */
  domiciliCalle: string;

  /** Número exterior del domicilio. */
  domiciliNumeroExterior: string;

  /** Número interior del domicilio. */
  domiciliNumeroInterior: string;

  /** Código LADA asociado al teléfono del destinatario. */
  domiciliLada: string;

  /** Número telefónico del destinatario. */
  domiciliTelefono: string;

  /** Correo electrónico del destinatario. */
  domiciliCorreoElectronioco: string;
}